// app/(protected)/vendas/pedidos-de-venda/types.ts
export interface PedidoVendaProps {
  codigo_pedido_omie: number;
  numero_pedido: string | null;
  sequencial: number | null;
  etapa: number | null;
  data_inclusao: string | null;
  hora_inclusao: string | null;
  data_previsao: string | null;
  data_faturamento: string | null;
  hora_faturamento: string | null;
  data_cancelamento: string | null;
  hora_cancelamento: string | null;
  data_encerramento: string | null;
  hora_encerramento: string | null;
  codigo_empresa: string;
  codigo_cliente: string;
  codigo_vendedor_omie: string;
  codigo_categoria: string | null;
  codigo_projeto: string | null;
  numero_contrato: string | null;
  obs_venda: string | null;
  valor_total_pedido: number | null;
  motivo_encerramento: string | null;
  usuario_encerramento: string | null;
  situacao: string | null;
  autorizado: boolean;
  denegado: boolean;
  faturado: boolean;
  cancelado: boolean;
  devolvido: boolean;
  devolucao_parcial: boolean;
  encerrado: boolean;
  created_by: string | null;
  updated_by: string | null;
  deleted_by: string | null;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
}